# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shadow Operator HQ API — unified creator-outreach + sales-pipeline CRM
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  fullName: string | null;
  /** @nullable */
  email: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface SignUpRequest {
  /** @minLength 1 */
  fullName: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LogInRequest {
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  password: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  new: "new",
  contacted: "contacted",
  qualified: "qualified",
  proposal: "proposal",
  won: "won",
  lost: "lost",
} as const;

export interface Lead {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  instagramHandle?: string | null;
  /** @nullable */
  niche?: string | null;
  /** @nullable */
  subNiche?: string | null;
  /** @nullable */
  estFollowers?: string | null;
  /** @nullable */
  profileBioSummary?: string | null;
  /** @nullable */
  monetizationGap?: string | null;
  priorityScore: number;
  dmStatus: string;
  /** @nullable */
  response?: string | null;
  callBooked: string;
  /** @nullable */
  notes?: string | null;
  status: LeadStatus;
  createdAt: string;
  updatedAt: string;
}

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  company?: string;
  email?: string;
  phone?: string;
  instagramHandle?: string;
  niche?: string;
  subNiche?: string;
  estFollowers?: string;
  profileBioSummary?: string;
  monetizationGap?: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  priorityScore?: number;
  dmStatus?: string;
  response?: string;
  callBooked?: string;
  notes?: string;
  status?: LeadStatus;
}

export interface LeadUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  instagramHandle?: string | null;
  /** @nullable */
  niche?: string | null;
  /** @nullable */
  subNiche?: string | null;
  /** @nullable */
  estFollowers?: string | null;
  /** @nullable */
  profileBioSummary?: string | null;
  /** @nullable */
  monetizationGap?: string | null;
  /**
   * @minimum 1
   * @maximum 10
   */
  priorityScore?: number;
  dmStatus?: string;
  /** @nullable */
  response?: string | null;
  callBooked?: string;
  /** @nullable */
  notes?: string | null;
  status?: LeadStatus;
}

export type FollowUpStatus =
  (typeof FollowUpStatus)[keyof typeof FollowUpStatus];

export const FollowUpStatus = {
  pending: "pending",
  completed: "completed",
} as const;

export interface FollowUp {
  id: number;
  leadId: number;
  userId: string;
  scheduledDate: string;
  status: FollowUpStatus;
  dayOffset: number;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  leadName?: string | null;
  /** @nullable */
  leadCompany?: string | null;
  /** @nullable */
  leadEmail?: string | null;
  /** @nullable */
  leadInstagramHandle?: string | null;
}

export interface DashboardFollowups {
  overdue: FollowUp[];
  today: FollowUp[];
  upcoming: FollowUp[];
}

export interface NicheCount {
  /** @nullable */
  niche: string | null;
  count: number;
}

export interface DmStatusCount {
  dmStatus: string;
  count: number;
}

export interface StatusCount {
  status: LeadStatus;
  count: number;
}

export interface LeadStats {
  totalLeads: number;
  callsBooked: number;
  avgPriorityScore: number;
  nicheBreakdown: NicheCount[];
  dmStatusBreakdown: DmStatusCount[];
  statusBreakdown: StatusCount[];
}

export type ListLeadsParams = {
  search?: string;
  niche?: string;
  dmStatus?: string;
  callBooked?: string;
  status?: LeadStatus;
  minPriority?: number;
  maxPriority?: number;
};