# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shadow Operator HQ API — unified creator-outreach + sales-pipeline CRM
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Authentication endpoints.
  - name: leads
    description: Lead management operations
  - name: followups
    description: Follow-up management operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'

  /auth/signup:
    post:
      tags: [Auth]
      operationId: signUp
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: Account created and session started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Email already registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /auth/login:
    post:
      tags: [Auth]
      operationId: logIn
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogInRequest'
      responses:
        '200':
          description: Logged in successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
        '401':
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /auth/logout:
    post:
      tags: [Auth]
      operationId: logOut
      summary: End the current session
      responses:
        '200':
          description: Logged out.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /leads/stats:
    get:
      operationId: getLeadStats
      tags: [leads]
      summary: Get lead statistics
      description: Returns aggregated stats - niche breakdown, DM status counts, pipeline status counts, priority distribution
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadStats"

  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: niche
          in: query
          schema:
            type: string
        - name: dmStatus
          in: query
          schema:
            type: string
        - name: callBooked
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/LeadStatus"
        - name: minPriority
          in: query
          schema:
            type: integer
        - name: maxPriority
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      description: Creates the lead and auto-generates its follow-up schedule (day 0, 3, 7, 11, 14)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Validation error

  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /leads/{id}/followups:
    get:
      operationId: getLeadFollowups
      tags: [followups]
      summary: Get follow-ups for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of follow-ups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowUp"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /followups:
    get:
      operationId: listFollowups
      tags: [followups]
      summary: List pending follow-ups for current user, grouped by urgency
      responses:
        "200":
          description: Follow-ups grouped by urgency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardFollowups"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /followups/{id}/complete:
    patch:
      operationId: completeFollowup
      tags: [followups]
      summary: Mark a follow-up as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Follow-up marked complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthUser:
      type: object
      required: [id, fullName, email]
      properties:
        id:
          type: string
        fullName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]

    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'

    SignUpRequest:
      type: object
      required: [fullName, email, password]
      properties:
        fullName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 8

    LogInRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 1

    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    LeadStatus:
      type: string
      enum:
        - new
        - contacted
        - qualified
        - proposal
        - won
        - lost

    Lead:
      type: object
      required:
        - id
        - userId
        - name
        - priorityScore
        - dmStatus
        - callBooked
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        company:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        instagramHandle:
          type: ["string", "null"]
        niche:
          type: ["string", "null"]
        subNiche:
          type: ["string", "null"]
        estFollowers:
          type: ["string", "null"]
        profileBioSummary:
          type: ["string", "null"]
        monetizationGap:
          type: ["string", "null"]
        priorityScore:
          type: integer
        dmStatus:
          type: string
        response:
          type: ["string", "null"]
        callBooked:
          type: string
        notes:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/LeadStatus"
        createdAt:
          type: string
        updatedAt:
          type: string

    LeadInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        company:
          type: string
        email:
          type: string
        phone:
          type: string
        instagramHandle:
          type: string
        niche:
          type: string
        subNiche:
          type: string
        estFollowers:
          type: string
        profileBioSummary:
          type: string
        monetizationGap:
          type: string
        priorityScore:
          type: integer
          minimum: 1
          maximum: 10
        dmStatus:
          type: string
        response:
          type: string
        callBooked:
          type: string
        notes:
          type: string
        status:
          $ref: "#/components/schemas/LeadStatus"

    LeadUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        company:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        instagramHandle:
          type: ["string", "null"]
        niche:
          type: ["string", "null"]
        subNiche:
          type: ["string", "null"]
        estFollowers:
          type: ["string", "null"]
        profileBioSummary:
          type: ["string", "null"]
        monetizationGap:
          type: ["string", "null"]
        priorityScore:
          type: integer
          minimum: 1
          maximum: 10
        dmStatus:
          type: string
        response:
          type: ["string", "null"]
        callBooked:
          type: string
        notes:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/LeadStatus"

    FollowUpStatus:
      type: string
      enum:
        - pending
        - completed

    FollowUp:
      type: object
      required:
        - id
        - leadId
        - userId
        - scheduledDate
        - status
        - dayOffset
      properties:
        id:
          type: integer
        leadId:
          type: integer
        userId:
          type: string
        scheduledDate:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/FollowUpStatus"
        dayOffset:
          type: integer
        completedAt:
          type: ["string", "null"]
          format: date-time
        leadName:
          type: ["string", "null"]
        leadCompany:
          type: ["string", "null"]
        leadEmail:
          type: ["string", "null"]
        leadInstagramHandle:
          type: ["string", "null"]

    DashboardFollowups:
      type: object
      required: [overdue, today, upcoming]
      properties:
        overdue:
          type: array
          items:
            $ref: "#/components/schemas/FollowUp"
        today:
          type: array
          items:
            $ref: "#/components/schemas/FollowUp"
        upcoming:
          type: array
          items:
            $ref: "#/components/schemas/FollowUp"

    NicheCount:
      type: object
      required: [niche, count]
      properties:
        niche:
          type: ["string", "null"]
        count:
          type: integer

    DmStatusCount:
      type: object
      required: [dmStatus, count]
      properties:
        dmStatus:
          type: string
        count:
          type: integer

    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          $ref: "#/components/schemas/LeadStatus"
        count:
          type: integer

    LeadStats:
      type: object
      required:
        - totalLeads
        - callsBooked
        - avgPriorityScore
        - nicheBreakdown
        - dmStatusBreakdown
        - statusBreakdown
      properties:
        totalLeads:
          type: integer
        callsBooked:
          type: integer
        avgPriorityScore:
          type: number
        nicheBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/NicheCount"
        dmStatusBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DmStatusCount"
        statusBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"