# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shadow Operator HQ API — unified creator-outreach + sales-pipeline CRM
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  /** @nullable */
  fullName: string | null;
  /** @nullable */
  email: string | null;
}