# lib/api-zod/src/generated/types/followUp.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shadow Operator HQ API — unified creator-outreach + sales-pipeline CRM
 * OpenAPI spec version: 0.1.0
 */
import type { FollowUpStatus } from "./followUpStatus";

export interface FollowUp {
  id: number;
  leadId: number;
  userId: string;
  scheduledDate: Date;
  status: FollowUpStatus;
  dayOffset: number;
  /** @nullable */
  completedAt?: Date | null;
  /** @nullable */
  leadName?: string | null;
  /** @nullable */
  leadCompany?: string | null;
  /** @nullable */
  leadEmail?: string | null;
  /** @nullable */
  leadInstagramHandle?: string | null;
}