# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shadow Operator HQ API — unified creator-outreach + sales-pipeline CRM
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from "./leadStatus";

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  company?: string;
  email?: string;
  phone?: string;
  instagramHandle?: string;
  niche?: string;
  subNiche?: string;
  estFollowers?: string;
  profileBioSummary?: string;
  monetizationGap?: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  priorityScore?: number;
  dmStatus?: string;
  response?: string;
  callBooked?: string;
  notes?: string;
  status?: LeadStatus;
}