# lib/api-zod/src/generated/types/leadStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shadow Operator HQ API — unified creator-outreach + sales-pipeline CRM
 * OpenAPI spec version: 0.1.0
 */
import type { DmStatusCount } from "./dmStatusCount";
import type { NicheCount } from "./nicheCount";
import type { StatusCount } from "./statusCount";

export interface LeadStats {
  totalLeads: number;
  callsBooked: number;
  avgPriorityScore: number;
  nicheBreakdown: NicheCount[];
  dmStatusBreakdown: DmStatusCount[];
  statusBreakdown: StatusCount[];
}