# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shadow Operator HQ API — unified creator-outreach + sales-pipeline CRM
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from "./leadStatus";

export interface LeadUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  instagramHandle?: string | null;
  /** @nullable */
  niche?: string | null;
  /** @nullable */
  subNiche?: string | null;
  /** @nullable */
  estFollowers?: string | null;
  /** @nullable */
  profileBioSummary?: string | null;
  /** @nullable */
  monetizationGap?: string | null;
  /**
   * @minimum 1
   * @maximum 10
   */
  priorityScore?: number;
  dmStatus?: string;
  /** @nullable */
  response?: string | null;
  callBooked?: string;
  /** @nullable */
  notes?: string | null;
  status?: LeadStatus;
}