# lib/api-zod/src/generated/types/logInRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shadow Operator HQ API — unified creator-outreach + sales-pipeline CRM
 * OpenAPI spec version: 0.1.0
 */

export interface LogInRequest {
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  password: string;
}