# lib/api-zod/src/generated/types/statusCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shadow Operator HQ API — unified creator-outreach + sales-pipeline CRM
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from "./leadStatus";

export interface StatusCount {
  status: LeadStatus;
  count: number;
}